R059: AGI bridge speaks CGI — text commands drive the whole 3D pipeline.
Agent protocol extensions (wb_agent_command):
- cgi-box/sphere/cylinder/cone <x y z size r g b> — add primitives
- cgi-key <obj> <t> <px py pz scale> [ease] — keyframe motion
- cgi-asset <kit> <model> <x y z> — stamp a Kenney/library GLB (with
  wb_assets_release ownership transfer fixing a cache-lifetime bug)
- cgi-list — enumerate kits/models
- cgi-render <out.mp4> [start dur] — wb_video_export_cgi end-to-end

test_agent_cgi = 9 checks: primitives, keys, render->ffprobe-verified
video, listing, asset stamping. Full gate green across all 11 test
binaries + selftest + E2E export.

+++ include/wbus/wbus.h
@@ -348,7 +348,7 @@ int  wb_video_export_codec(wb_session *s, wb_engine *e,
 
 /* Write 16-bit PCM WAV from interleaved stereo samples (export utility). */
 int  wb_wav_write_pcm16(const char *path, const wb_sample *data, uint32_t frames,
-                        uint32_t channels, uint32_t sample_rate);
+                        uint8_t channels, uint32_t sample_rate);
 
 /* Legacy H.264 wrapper (delegates to wb_video_export_codec). */
 int  wb_video_export(wb_session *s, wb_engine *e,
